refactor: OpenAPI 처리 로직을 build.gradle.kts로 이동 및 https 적용
