--- backend/buildSrc/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    `kotlin-dsl`
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation(gradleApi())
-    implementation(localGroovy())
-}
